Create addresses module
